// frontend/src/types/auth.types.ts
export interface Sede {
  id: number;
  name: string;
  address: string;
  phone: string;
}
// ...
export interface Pagination {
  total: number;
  page: number;
  page_size: number;
  total_pages: number;
}
// ...
export interface SedeDetail extends Sede {
  is_active: boolean;
  created_at: string;
  updated_at: string;
  user_count: number;
}